# openapi/definicion.yaml
openapi: 3.1.0
info:
  title: CatalogIA API
  description: FastAPI backend for CatalogIA platform with AI recommendations
  version: 1.0.0
  contact:
    name: Equipo 22
    email: 'equipo22@tec.mx'
    url: http://localhost:8000
servers:
- url: http://localhost:8000
  description: Local development server
paths:
...
  /api/v1/products/:
    post:
      tags:
      - products
      summary: Create Product
      description: "Crear un nuevo producto en el catálogo.\n\nGenera automáticamente\
        \ el embedding semántico del producto.\n\nArgs:\n    - `db`: Sesión de base\
        \ de datos \n    - `product_in`: Datos del producto a crear\n\nReturns:\n\
        \    `ProductResponse`: Producto creado con todos sus datos\n\nRaises:\n \
        \   `HTTPException`: 400 si ya existe un producto con el mismo EAN"
      operationId: create_product_api_v1_products__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse-Output'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - products
      summary: List Products
      description: "Listar productos con filtros opcionales y paginación.\n\nArgs:\n\
        \    - `skip`: Número de productos a omitir (para paginación)\n    - `limit`:\
        \ Número máximo de productos a retornar\n    - `category`: Filtro por categoría\n\
        \    - `min_price`: Precio mínimo\n    - `max_price`: Precio máximo\n\nReturns:\n\
        \    `ProductList`: Lista paginada de productos con metadatos"
      operationId: list_products_api_v1_products__get
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Número de productos a saltar para paginación
          default: 0
          title: Skip
        description: Número de productos a saltar para paginación
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          description: Límite de productos por página
          default: 100
          title: Limit
        description: Límite de productos por página
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filtrar por categoría
          title: Category
        description: Filtrar por categoría
      - name: min_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            minimum: 0
          - type: 'null'
          description: Precio mínimo
          title: Min Price
        description: Precio mínimo
      - name: max_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            minimum: 0
          - type: 'null'
          description: Precio máximo
          title: Max Price
        description: Precio máximo
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/products/search/semantic:
    get:
      tags:
      - products
      summary: Semantic Search Products
      description: "Búsqueda semántica de productos usando IA.\n\nEsta búsqueda utiliza\
        \ embeddings y similitud semántica para encontrar\nproductos relevantes basándose\
        \ en el SIGNIFICADO del texto, no solo\nen coincidencias exactas de palabras.\n\
        \nArgs:\n    - `q`: Texto de búsqueda (ej: \"laptop para trabajar desde casa\"\
        )\n    - `limit`: Número máximo de resultados a retornar\n    - `category`:\
        \ Filtro opcional por categoría\n    - `min_price`: Precio mínimo del filtro\n\
        \    - `max_price`: Precio máximo del filtro\n    - `min_similarity`: Umbral\
        \ mínimo de similitud (0.3 recomendado, 0.5 restrictivo)\n\nReturns:\n   \
        \ `SemanticSearchResult`: Productos ordenados por relevancia con scores de\
        \ similitud"
      operationId: semantic_search_products_api_v1_products_search_semantic_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          description: Texto de búsqueda
          title: Q
        description: Texto de búsqueda
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Número de resultados
          default: 10
          title: Limit
        description: Número de resultados
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filtrar por categoría
          title: Category
        description: Filtrar por categoría
      - name: min_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            minimum: 0
          - type: 'null'
          description: Precio mínimo
          title: Min Price
        description: Precio mínimo
      - name: max_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            minimum: 0
          - type: 'null'
          description: Precio máximo
          title: Max Price
        description: Precio máximo
      - name: min_similarity
        in: query
        required: false
        schema:
          type: number
          maximum: 1
          minimum: 0
          description: Umbral mínimo de similitud (0-1)
          default: 0.3
          title: Min Similarity
        description: Umbral mínimo de similitud (0-1)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticSearchResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/products/{product_id}:
    get:
      tags:
      - products
      summary: Get Product
      description: "Obtener un producto específico por su ID.\n\nArgs:\n    - `product_id`:\
        \ ID del producto a buscar\n\nReturns:\n    `ProductResponse`: Datos completos\
        \ del producto\n\nRaises:\n    `HTTPException`: 404 si el producto no existe"
      operationId: get_product_api_v1_products__product_id__get
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse-Output'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - products
      summary: Update Product
      description: "Actualizar un producto existente.\n\nRegenera automáticamente\
        \ el embedding si se modifican campos relevantes\n(título, descripción, categoría,\
        \ subcategoría, o características).\n\nArgs:\n    - `product_id`: ID del producto\
        \ a actualizar\n    - `product_in`: Datos a actualizar (solo campos presentes\
        \ se modifican)\n\nReturns:\n    `ProductResponse`: Producto actualizado\n\
        \nRaises:\n    `HTTPException`: 404 si el producto no existe\n    `HTTPException`:\
        \ 400 si el nuevo EAN ya está en uso"
      operationId: update_product_api_v1_products__product_id__put
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
          title: Product Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse-Output'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - products
      summary: Delete Product
      description: "Eliminar un producto del catálogo.\n\nArgs:\n    - `product_id`:\
        \ ID del producto a eliminar\n\nReturns:\n    `dict`: Mensaje de confirmación\n\
        \nRaises:\n    `HTTPException`: 404 si el producto no existe"
      operationId: delete_product_api_v1_products__product_id__delete
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
      description: Error de validación HTTP que contiene una lista de errores de validación
        detallados.
...
    ProductCreate:
      properties:
        ean:
          type: string
          maxLength: 13
          minLength: 13
          title: Ean
        title:
          type: string
          title: Title
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        product_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Description
        category:
          type: string
          title: Category
        price:
          anyOf:
          - type: number
            exclusiveMinimum: 0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Price
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        discount:
          anyOf:
          - type: number
            maximum: 100
            minimum: 0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Discount
          default: 0
        rating:
          anyOf:
          - type: number
            maximum: 5
            minimum: 0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Rating
          default: 0
        stock:
          type: integer
          minimum: 0
          title: Stock
        sponsored:
          type: boolean
          title: Sponsored
          default: false
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
      type: object
      required:
      - ean
      - title
      - category
      - price
      - stock
      title: ProductCreate
      description: Esquema para crear un nuevo producto en el catálogo. Requiere EAN,
        título, categoría, precio y stock. Genera automáticamente el embedding semántico.
    ProductList:
      properties:
        products:
          items:
            $ref: '#/components/schemas/ProductResponse-Output'
          type: array
          title: Products
        total:
          type: integer
          title: Total
        skip:
          type: integer
          title: Skip
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - products
      - total
      - skip
      - limit
      title: ProductList
      description: Respuesta paginada con lista de productos, incluyendo el total
        de productos disponibles y los parámetros de paginación aplicados.
...
    ProductResponse-Output:
      properties:
        ean:
          type: string
          maxLength: 13
          minLength: 13
          title: Ean
        title:
          type: string
          title: Title
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        product_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Description
        category:
          type: string
          title: Category
        price:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Price
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        discount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Discount
          default: 0
        rating:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Rating
          default: 0
        stock:
          type: integer
          minimum: 0
          title: Stock
        sponsored:
          type: boolean
          title: Sponsored
          default: false
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - ean
      - title
      - category
      - price
      - stock
      - id
      - created_at
      - updated_at
      title: ProductResponse
      description: Esquema de respuesta con la información completa de un producto,
        con campos numéricos serializados como strings para precisión decimal.
    ProductUpdate:
      properties:
        ean:
          anyOf:
          - type: string
            maxLength: 13
            minLength: 13
          - type: 'null'
          title: Ean
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        product_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Description
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        price:
          anyOf:
          - type: number
            exclusiveMinimum: 0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Price
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        discount:
          anyOf:
          - type: number
            maximum: 100
            minimum: 0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Discount
        rating:
          anyOf:
          - type: number
            maximum: 5
            minimum: 0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Rating
        stock:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
          title: Stock
        sponsored:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Sponsored
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
      type: object
      title: ProductUpdate
      description: Esquema para actualizar un producto existente. Todos los campos
        son opcionales. Si se modifican campos relevantes, se regenera automáticamente
        el embedding semántico.
...
    ProductWithScore-Output:
      properties:
        ean:
          type: string
          maxLength: 13
          minLength: 13
          title: Ean
        title:
          type: string
          title: Title
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        product_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Description
        category:
          type: string
          title: Category
        price:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Price
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        discount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Discount
          default: 0
        rating:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Rating
          default: 0
        stock:
          type: integer
          minimum: 0
          title: Stock
        sponsored:
          type: boolean
          title: Sponsored
          default: false
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        similarity_score:
          type: number
          maximum: 1
          minimum: 0
          title: Similarity Score
          description: Score de similitud (0-1, mayor es más similar)
      type: object
      required:
      - ean
      - title
      - category
      - price
      - stock
      - id
      - created_at
      - updated_at
      - similarity_score
      title: ProductWithScore
      description: Producto con score de similitud semántica.
    SemanticSearchResult:
      properties:
        products:
          items:
            $ref: '#/components/schemas/ProductWithScore-Output'
          type: array
          title: Products
        total:
          type: integer
          title: Total
        skip:
          type: integer
          title: Skip
        limit:
          type: integer
          title: Limit
        min_similarity:
          anyOf:
          - type: number
          - type: 'null'
          title: Min Similarity
          description: Umbral mínimo de similitud aplicado
      type: object
      required:
      - products
      - total
      - skip
      - limit
      title: SemanticSearchResult
      description: Resultado de búsqueda semántica con scores de similitud.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
      description: Detalle de un error de validación individual que incluye la ubicación
        del error, el mensaje descriptivo y el tipo de error.
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer